backend/requirements.txt
numpy>=1.26.0,<3.0.0
pandas>=2.2.0,<3.0.0
matplotlib>=3.8.0,<4.0.0
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
statsmodels>=0.14.1,<1.0.0
psycopg[binary]>=3.2.0,<4.0.0
tqdm>=4.66.0,<5.0.0
fastapi>=0.115.0,<1.0.0
uvicorn>=0.30.0,<1.0.0
streamlit>=1.40.0,<2.0.0
plotly>=5.24.0,<6.0.0
APScheduler>=3.10.4,<4.0.0
openpyxl>=3.1.0,<4.0.0
